Use DefaultBatchedGrandProduct for timestamp validity, read-write memory and bytecode
